Preliminary version of the SAX parser for specifying constraints on input files
git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@804 0ae5e6b2-d358-4f09-a895-f82f13dd62a4

+++ src/joshua/decoder/segment_file/ConstraintRule.java
@@ -46,5 +46,5 @@ public interface ConstraintRule {
 	/**
 	 * Return the value for the i'th grammar feature.
 	 */
-	float feature(int i);
+	double feature(int i);
 }
